linkloop: Print statistics in case of errors
Sometimes it's handy to have the packets statistics printed at the
end of the ping procedure. For instance it's good to know how many
packets actually transmitted, what part of them actually successfully
retrieved, which one of them never got back and which one got back
with errors. This commit provides the corresponding functionality.

Signed-off-by: Serge Semin <fancer.lancer@gmail.com>