linkloop: Detect rogue responder only if the frame really substituted
Indeed a situation when destination MAC address of the source frame doesn't
matches the response frame source isn't erroneous in case if the system has
two loopbacked interfaces, so the ping frame just gets back via the second
interface. In this case the source MAC addresses of the sent and retrieved
frames match.

Signed-off-by: Serge Semin <fancer.lancer@gmail.com>